% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_helpers.R
\name{pr_request_url}
\alias{pr_request_url}
\title{Build a pull/merge request URL from a git remote}
\usage{
pr_request_url(remote_url, source_branch, target_branch)
}
\arguments{
\item{remote_url}{A git remote URL.}

\item{source_branch}{The source branch name.}

\item{target_branch}{The target branch name.}
}
\value{
A URL string that can be used to open a new pull request (GitHub) or
merge request (GitLab).
}
\description{
Build a pull/merge request URL from a git remote
}
\examples{
pr_request_url(
  "https://github.com/OWNER/REPO.git",
  source_branch = "feature/new-thing",
  target_branch = "main"
)
}
