% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.helpers-package.R
\docType{package}
\name{pr.helpers-package}
\alias{pr.helpers}
\alias{pr.helpers-package}
\title{pr.helpers: Helpers for Pull and Merge Request Git Workflows}
\description{
Provides helper functions for common pull request and merge request workflows from a local Git repository. Includes utilities for creating and switching branches, fetching requests by number, generating request URLs, and syncing with the default branch across GitHub and GitLab remotes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/d-morrison/pr.helpers}
  \item Report bugs at \url{https://github.com/d-morrison/pr.helpers/issues}
}

}
\author{
\strong{Maintainer}: David Morrison \email{2474437+d-morrison@users.noreply.github.com}

}
