% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.helpers-package.R
\docType{package}
\name{pr.helpers-package}
\alias{pr.helpers}
\alias{pr.helpers-package}
\title{pr.helpers: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UCD-SERG/pr.helpers}
  \item \url{https://ucd-serg.github.io/pr.helpers/}
  \item Report bugs at \url{https://github.com/UCD-SERG/pr.helpers/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
